Updated these migrations in order to communicate better

+++ database/migrations/2023_09_14_185642_create_user_infos_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('user_infos', function (Blueprint $table) {
             $table->id();
             $table->foreignId('user_id')->constrained()->cascadeOnDelete();
-            $table->unsignedSmallInteger('sa_id')->unique();
+            $table->bigInteger('sa_id')->unique();
             $table->string('mobile_number', 20)->unique();
             $table->timestamp('birth_date');
             $table->string('language');
